requirements.txt
amqp==5.1.1
APScheduler==3.10.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudant==2.15.0
cryptography==40.0.2
dimagi-utils==1.2.3.2
et-xmlfile==1.1.0
gevent==22.10.2
greenlet==2.0.2
http-parser==0.9.0
idna==3.4
jsonobject==2.1.0
jsonobject-couchdbkit==1.0.1
kombu==5.2.4
linecache2==1.0.0
mock==5.0.2
msgpack-python==0.5.6
openpyxl==3.1.2
peewee==3.16.2
Pillow==9.5.0
pillowtop==0.1.5
pip==21.3.1
prompt-toolkit==3.0.38
pycparser==2.21
PyMySQL==1.0.3
pypng==0.20220715.0
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
qrcode==7.4.2
rawes==0.5.5
redis==4.5.4
requests==2.30.0
restkit==4.2.2
setuptools==60.2.0
simplejson==3.19.1
six==1.16.0
socketpool==0.5.3
traceback2==1.4.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
unittest2==1.1.0
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
wheel==0.37.1
zope.event==4.6
zope.interface==6.0

Flask~=2.3.2